move initInfinex to remove wagmi sign out bug
