+++ pages/_app.page.tsx
@@ -2,6 +2,7 @@ import '/public/fonts/inter/inter.css';
 import '/src/styles/variables.css';
 
 import { CacheProvider, EmotionCache } from '@emotion/react';
+import { InfinexProvider } from '@infinex/connect-sdk';
 import { NoSsr } from '@mui/material';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
@@ -20,14 +21,13 @@ import { ModalContextProvider } from 'src/hooks/useModal';
 import { Web3ContextProvider } from 'src/libs/web3-data-provider/Web3Provider';
 import { useRootStore } from 'src/store/root';
 import { SharedDependenciesProvider } from 'src/ui-config/SharedDependenciesProvider';
-import { wagmiConfig } from 'src/ui-config/wagmiConfig';
+import { infinexCore, wagmiConfig } from 'src/ui-config/wagmiConfig';
 import { WagmiProvider } from 'wagmi';
 import { useShallow } from 'zustand/shallow';
 
 import createEmotionCache from '../src/createEmotionCache';
 import { AppGlobalStyles } from '../src/layouts/AppGlobalStyles';
 import { LanguageProvider } from '../src/libs/LanguageProvider';
-import { InfinexConnectConfig, InfinexProvider, initInfinex } from '@infinex/connect-sdk';
 
 const SwitchModal = dynamic(() =>
   import('src/components/transactions/Switch/SwitchModal').then((module) => module.SwitchModal)
@@ -111,19 +111,6 @@ export default function MyApp(props: MyAppProps) {
         },
       })
   );
-
-  const infinexConfig: InfinexConnectConfig = {
-    appKey: 'example',
-    debug: {
-      showLogs: true,
-      iframeOptions: {
-        shown: true,
-        width: '750',
-        height: '405',
-      },
-    },
-  };
-  const infinexCore = initInfinex(infinexConfig);
 
   const MIXPANEL_TOKEN = process.env.NEXT_PUBLIC_MIXPANEL;
   useEffect(() => {
